Use a more positive title
The mailing list has been discussing a change of title to a more
neutral form. The words in this title have not been approved by the
group but are better than we had previously. The title may yet
change. Many people were involved in tweaking to get to this version.