Remove section 2.1.4. Informal Variants
This section largely argues against poor software development
practices which is not the focus of the paper. The parts which
deal with how FITS needlessly makes it easy (arguably promotes)
these bad practices by its lack of good validation has been kept
along with the issue that informal variants themselves pose a problem
for the user community. This content has been shifted and encorporated
in an earlier commit to section 2.1.2, Inadequate declaration of
content meaning.