First stab at a discussion section.
Going with a 'lessions learned' approach, as suggested in the
title. So far, I've gotten 4 lessions, of which in this commit,
2 are written up (although the text is still a bit rough). As
part of this commit, I've pulled all of the 'suggestion about
what FITS community should do' text from the proceeding section 2,
and placed what seemed to fit in the discussion. Some of the text
is redded out, which means its simply notes for a section, not
ready to be read critically yet.