In section 2.3.1, highlight hierarchical models
text. It doesnt appear to add to the argument. We are not
particularly stating problems with these models. I imagine
problems in the implementation do exist, but we need to
state them (such as, and Im guessing, "HIERARCH model
results in scattering metadata throughout the file and is
not particularly easy to reconstruct without the help of
specialized software")