red out compression section. It is problematic
for a variety of reasons. First, it winds up a subset
of readers..FITS compression is considered a thing FITS
does right. Second, its a summary of the results of a
companion paper, so it seems to be stealing the wind from
the conclusions there. Last, I can see a valid and strong
argument for making compression separate from the standard.
Only in the case of dramatic compression performance would
it seem merited to bring in a solution to the standard. This
last point seems to be a touchstone of controversy for some
readers, and makes the second a distraction from the greater
flow of the document.